Reshuffle the GIS layer mapping code so the mapping gets saved + code cleanup
