+++ src/main/java/com/socrata/datasync/job/GISJob.java
@@ -1,20 +1,20 @@
 package com.socrata.datasync.job;
 
 import com.socrata.api.SodaImporter;
-import com.socrata.datasync.PublishMethod;
-import com.socrata.datasync.SMTPMailer;
-import com.socrata.datasync.SocrataConnectionInfo;
-import com.socrata.datasync.Utils;
+import com.socrata.datasync.*;
 import com.socrata.datasync.config.CommandLineOptions;
 import com.socrata.datasync.config.controlfile.ControlFile;
 import com.socrata.datasync.config.userpreferences.UserPreferences;
 import com.socrata.datasync.config.userpreferences.UserPreferencesJava;
 import com.socrata.datasync.publishers.GISPublisher;
 import com.socrata.datasync.validation.GISJobValidity;
+import com.socrata.model.importer.Dataset;
 import com.socrata.model.importer.GeoDataset;
 
 
 import org.apache.commons.cli.CommandLine;
+import org.apache.commons.io.FilenameUtils;
+import org.apache.http.HttpException;
 import org.codehaus.jackson.annotate.JsonIgnoreProperties;
 import org.codehaus.jackson.annotate.JsonProperty;
 import org.codehaus.jackson.map.DeserializationConfig;
@@ -22,9 +22,11 @@ import org.codehaus.jackson.map.ObjectMapper;
 import org.codehaus.jackson.map.annotate.JsonSerialize;
 
 import java.io.*;
-import java.util.HashMap;
-import java.util.Map;
+import java.net.URISyntaxException;
+import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipFile;
 
 @JsonIgnoreProperties(ignoreUnknown=true)
 @JsonSerialize(include=JsonSerialize.Inclusion.NON_NULL)
@@ -171,7 +173,6 @@ public class GISJob extends Job {
      * @param cmd the commandLine object constructed from the user's options
      */
     public void configure(CommandLine cmd) {
-        CommandLineOptions options = new CommandLineOptions();
         String method = cmd.getOptionValue(CommandLineOptions.PUBLISH_METHOD_FLAG);
         setDatasetID(cmd.getOptionValue(CommandLineOptions.DATASET_ID_FLAG));
         setFileToPublish(cmd.getOptionValue(CommandLineOptions.FILE_TO_PUBLISH_FLAG));
@@ -195,12 +196,11 @@ public class GISJob extends Job {
      */
     public JobStatus run() {
         SocrataConnectionInfo connectionInfo = userPrefs.getConnectionInfo();
-        GeoDataset result = null;
         String publishExceptions = "";
         JobStatus runStatus = JobStatus.SUCCESS;
 
         JobStatus validationStatus = GISJobValidity.validateJobParams(connectionInfo, this);
-        JobStatus layerMappingStatus = GISJobValidity.initializeLayerMapping(userPrefs, getDatasetID(), this);
+        JobStatus layerMappingStatus = GISJobValidity.validateLayerMapping(this);
         if (validationStatus.isError()) {
             runStatus = validationStatus;
         } else if (layerMappingStatus.isError()) {
@@ -208,10 +208,6 @@ public class GISJob extends Job {
         } else {
             try {
                 File fileToPublishFile = new File(fileToPublish);
-                // attach a requestId to all Producer API calls (for error tracking purposes)
-                String jobRequestId = Utils.generateRequestId();
-                final SodaImporter importer = SodaImporter.newImporter(connectionInfo.getUrl(), connectionInfo.getUser(), connectionInfo.getPassword(), connectionInfo.getToken());
-
                 if (publishMethod == PublishMethod.replace) {
                     runStatus = GISPublisher.replaceGeo(fileToPublishFile, connectionInfo, datasetID, userPrefs);
                 } else {
@@ -231,6 +227,72 @@ public class GISJob extends Job {
 
         emailAdmin(runStatus);
         return runStatus;
+    }
+
+    /**
+     * This method attempts to map layers in the existing dataset to layers found in a shapefile,
+     * so that we replace existing layers where possible instead of creating new ones and changing
+     * the 4x4 / API endpoint of the dataset.
+     */
+    public void initializeLayerMapping() throws URISyntaxException, IOException, HttpException {
+        String fileToPublish = getFileToPublish();
+        String fileExtension = FilenameUtils.getExtension(fileToPublish);
+
+        if (fileExtension.equals(GISJobValidity.ZIP_EXT)) {
+            GeoDataset dataset = DatasetUtils.getDatasetInfo(userPrefs, getDatasetID(), GeoDataset.class);
+
+            // Get map of existing layer UIDs/names by looking at child views
+            Map<String, String> existingLayers = getLayerListFromExistingDataset(userPrefs, dataset);
+            // Get list of layer names in file by looking at .shp file names inside the zip
+            List<String> fileLayers = getLayerListFromShapefile(fileToPublish);
+
+            // Make a best effort to match file layers to existing layers by name
+            for (String fileLayer : fileLayers) {
+                if (existingLayers.containsKey(fileLayer)) {
+                    getLayerMap().put(fileLayer, existingLayers.get(fileLayer));
+                }
+            }
+
+            for (String key : getLayerMap().keySet()) {
+                System.out.println(key + " : " + getLayerMap().get(key));
+            }
+        } else {
+            setLayerMap(new HashMap<String, String>());
+        }
+    }
+
+    private static Map<String, String> getLayerListFromExistingDataset(UserPreferences userPrefs, GeoDataset dataset) {
+        List<String> existingLayersUids = dataset.getChildViews();
+        Map<String, String> existingLayerInfo = new HashMap<>();
+
+        for (String uid : existingLayersUids) {
+            try {
+                Dataset child = DatasetUtils.getDatasetInfo(userPrefs, uid, Dataset.class);
+                existingLayerInfo.put(child.getName(), uid);
+            } catch (Exception e) {
+                // there’s no way for the client to recover,
+                // so a checked exception is not necessary
+                throw new RuntimeException(e);
+            }
+        }
+
+        return existingLayerInfo;
+    }
+
+    private static List<String> getLayerListFromShapefile(String filePath) throws IOException {
+        try (ZipFile zipFile = new ZipFile(filePath)) {
+            List<String> layers = new ArrayList<>();
+            Enumeration entries = zipFile.entries();
+            while (entries.hasMoreElements()) {
+                ZipEntry entry = (ZipEntry) entries.nextElement();
+                if (FilenameUtils.getExtension(entry.getName()).equals("shp")) {
+                    String layerName = FilenameUtils.getBaseName(entry.getName());
+                    layers.add(layerName);
+                }
+            }
+
+            return layers;
+        }
     }
 
     private void sendErrorNotificationEmail(final String adminEmail,
@@ -258,14 +320,6 @@ public class GISJob extends Job {
         }
     }
 
-    private String logRunResults(JobStatus runStatus, GeoDataset result) {
-        String logDatasetID = userPrefs.getLogDatasetID();
-        String logPublishingErrorMessage = null;
-        SocrataConnectionInfo connectionInfo = userPrefs.getConnectionInfo();
-
-        return logPublishingErrorMessage;
-    }
-
     private void emailAdmin(JobStatus status) {
         String adminEmail = userPrefs.getAdminEmail();
         String logDatasetID = userPrefs.getLogDatasetID();
@@ -282,37 +336,4 @@ public class GISJob extends Job {
             super(msg);
         }
     }
-
-    /**
-     * This allows backward compatability with DataSync 0.1 .sij file format
-     *
-     * @param pathToFile .sij file that uses old serialization format (Java native)
-     * @throws IOException
-     */
-    private void loadOldSijFile(String pathToFile) throws IOException {
-        try {
-            InputStream file = new FileInputStream(pathToFile);
-            InputStream buffer = new BufferedInputStream(file);
-            ObjectInput input = new ObjectInputStream (buffer);
-            try{
-                com.socrata.datasync.IntegrationJob loadedJobOld =
-                    (com.socrata.datasync.IntegrationJob) input.readObject();
-
-                setDatasetID(loadedJobOld.getDatasetID());
-                setFileToPublish(loadedJobOld.getFileToPublish());
-                setPublishMethod(loadedJobOld.getPublishMethod());
-                setPathToSavedFile(pathToFile);
-            }
-            finally{
-                input.close();
-            }
-        } catch(Exception e) {
-            e.printStackTrace();
-
-            throw new IOException(e);
-        }
-    }
-
-
-
 }
